requirement.txt
flask
mysql-connector-python
flask-bcrypt
bcrypt==5.0.0
blinker==1.9.0
click==8.3.1
Flask==3.1.2
Flask-Bcrypt==1.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.5.0
Werkzeug==3.1.3
